requirements.txt
chardet==4.0.0
cx-Oracle==8.1.0
datefinder==0.7.3
idna==2.10
jdcal==1.4.1
python-dateutil==2.8.2
python-dotenv==0.15.0
pytz==2023.3
six==1.16.0
StrEnum==0.4.10
tzdata==2023.3
